Moved classes involved in the B-Rep for level geometry generation into a separate package -- seems to broadly make sense but can keep an eye as things develop
