+++ src/engine/brep/LoopSet.java
@@ -1,4 +1,4 @@
-package engine;
+package engine.brep;
 
 import java.util.ArrayList;
