+++ test/engine/brep/LoopTest.java
@@ -1,5 +1,6 @@
-package engine;
+package engine.brep;
 
+import engine.XY;
 import org.junit.Test;
 
 import java.util.ArrayList;
